reshuffled the entire project structure for more accessibility
