--- patient_data.csv
@@ -1,2 +1,0 @@
-First Name,Last Name,Email,Date of Birth,HIV Positive,Diagnosis Date,On ART,ART Start Date,Country Code
-Patient,GoodLuck,patient@example.com,1998-12-08,true,2021-01-27,false,,rw
