+++ resources/downloads/patient_data.csv
@@ -1,0 +1,4 @@
+First Name,Last Name,Email,Date of Birth,HIV Positive,Diagnosis Date,On ART,ART Start Date,Country Code
+Patient,GoodLuck,patient@example.com,1998-12-08,true,2021-01-27,false,,rw
+Billy One,Fon,billy@example.com,1985-10-12,true,2000-07-21,true,2003-01-27,ke
+Violette,Uwa,violette@example.com,1990-01-10,true,2000-01-01,true,2004-01-01,RW
