--- src/UserManager/models/UserRole.java
@@ -1,6 +1,0 @@
-package UserManager.models;
-
-public enum UserRole {
-    ADMIN,
-    PATIENT
-}
