+++ src/accounts/models/Admin.java
@@ -1,0 +1,15 @@
+
+package accounts.models;
+
+import utils.enums.UserRole;
+
+public class Admin extends User {
+    public Admin(String firstName, String lastName, String email, String password) {
+        super(firstName, lastName, email, password);
+    }
+
+    @Override
+    public UserRole getRole() {
+        return UserRole.ADMIN;
+    }
+}
