+++ src/config.properties
@@ -1,0 +1,8 @@
+usr.store=res/data/user-store.txt
+le.path=res/data/life-expectancy.csv
+
+sr.script=res/scripts/survival-rate.sh
+usr.script=res/scripts/user-manager.sh
+auth.script=res/scripts/auth.sh
+register.script=res/scripts/register.sh
+
