+++ src/core/BashConnect.java
@@ -1,0 +1,39 @@
+package core;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.util.logging.Logger;
+
+import accounts.services.UserManagementService;
+
+public class BashConnect {
+
+    private static final Logger logger = Logger.getLogger(UserManagementService.class.getName());
+
+    /**
+     * @return Returns from bash command
+     */
+    public static String run(String... command) throws IOException, InterruptedException {
+
+        ProcessBuilder processBuilder = new ProcessBuilder(command);
+        processBuilder.redirectErrorStream(true);
+        Process process = processBuilder.start();
+
+        StringBuilder output = new StringBuilder();
+        try (BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()))) {
+            String line;
+            while ((line = reader.readLine()) != null) {
+                output.append(line).append("\n");
+            }
+        }
+
+        int exitCode = process.waitFor();
+        if (exitCode != 0) {
+            logger.severe("Bash command execution failed" + output.toString());
+        }
+
+        return output.toString().trim();
+    }
+
+}
