+++ src/datacompute/models/PatientStatistic.java
@@ -1,0 +1,9 @@
+package datacompute.models;
+
+public class PatientStatistic {
+    private int patientId;
+    private String patientName;
+    private int age;
+    private String gender;
+    private String race;
+}
