--- src/helpers/ExportType.java
@@ -1,6 +1,0 @@
-package helpers;
-
-public enum ExportType {
-    PATIENT_INFO,
-    PATIENT_STATS
-}
