--- src/statistics/models/PatientStatistic.java
@@ -1,10 +1,0 @@
-package statistics.models;
-
-
-public class PatientStatistic {
-    private int patientId;
-    private String patientName;
-    private int age;
-    private String gender;
-    private String race;
-}
