--- src/statistics/services/StatServices.java
@@ -1,5 +1,0 @@
-package statistics.services;
-
-public class StatServices {
-
-}
