+++ src/utils/enums/ExportType.java
@@ -1,0 +1,6 @@
+package utils.enums;
+
+public enum ExportType {
+    PATIENT_INFO,
+    PATIENT_STATS
+}
