+++ src/utils/enums/UserRole.java
@@ -1,0 +1,6 @@
+package utils.enums;
+
+public enum UserRole {
+    ADMIN,
+    PATIENT
+}
