+++ src/utils/validators/Validator.java
@@ -1,0 +1,7 @@
+package utils.validators;
+
+public interface Validator {
+    boolean validate(String value);
+
+    String getErrorMessage();
+}
