--- user-store.txt
@@ -1,2 +1,0 @@
-patient@example.com,ae2c252f-4e91-4aba-9ea8-2c01088d3a46,Patient,GoodLuck,d4587ea9ead060c13fd994f21ecfa7926272a78854a2c20136b10a3c9e53e71e,patient,1998-12-08,true,2021-01-27,false,,rw
-admin@example.com,e9eefd12-70a5-45b3-b29a-27f25d76fe0e,Admin,Nyungwe,240be518fabd2724ddb6f04eeb1da5967448d7e831c08c8fa822809f74c720a9,admin
